fix: Move Landscape Map to landscape components
