+++ src/landscape/components/LandscapeForm/BoundaryStep/OptionAddPin.js
@@ -4,8 +4,8 @@ import { Paper, Typography } from '@mui/material';
 import PageHeader from 'layout/PageHeader';
 import DrawControls from 'gis/components/DrawControls';
 import { OPTION_BOUNDARY_CHOICES } from '.';
+import BaseMap from '../../LandscapeMap';
 import Actions from '../Actions';
-import BaseMap from './BaseMap';
 
 const OptionAddPin = props => {
   const { t } = useTranslation();
